refactor: change IDs from Integer to Long
- Updated database schema from INT to BIGINT
- Refactored all entities, repositories, and services to use Long
- Updated Controllers, DTOs, and Tests for type compability

+++ src/main/java/com/eduribeiro8/LilMarket/dto/BatchLossReportRequestDTO.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
 @Schema(description = "Requisição para reportar perda (danos/estrago) de um lote")
 public record BatchLossReportRequestDTO(
         @Schema(description = "Identificador do lote", example = "7", requiredMode = Schema.RequiredMode.REQUIRED)
-        @NotNull Integer batchId,
+        @NotNull Long batchId,
         @Schema(description = "Quantidade perdida", example = "2", requiredMode = Schema.RequiredMode.REQUIRED)
         @Positive BigDecimal quantity,
         @Schema(description = "Motivo da perda", example = "Quebrado durante transporte", requiredMode = Schema.RequiredMode.REQUIRED)
