+++ src/main/java/com/eduribeiro8/LilMarket/dto/SaleItemRequestDTO.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 @Schema(description = "Item incluído em uma requisição de venda")
 public record SaleItemRequestDTO(
         @Schema(description = "Identificador do produto", example = "10", requiredMode = Schema.RequiredMode.REQUIRED)
-        @NotNull @Positive Integer productId,
+        @NotNull @Positive Long productId,
         @Schema(description = "Quantidade do produto", example = "2", requiredMode = Schema.RequiredMode.REQUIRED)
         @NotNull @Positive BigDecimal quantity
 ) {}
