+++ src/main/java/com/eduribeiro8/LilMarket/dto/SaleItemResponseDTO.java
@@ -6,7 +6,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 @Schema(description = "Item retornado como parte de uma venda")
 public record SaleItemResponseDTO(
         @Schema(description = "Identificador do produto", example = "10")
-        Integer productId,
+        Long productId,
         @Schema(description = "Nome do produto", example = "Leite integral")
         String productName,
         @Schema(description = "Quantidade vendida", example = "2")
@@ -16,5 +16,5 @@ public record SaleItemResponseDTO(
         @Schema(description = "Subtotal deste item", example = "6.98")
         BigDecimal subtotal,
         @Schema(description = "Identificador do lote associado, quando aplicável", example = "7")
-        Integer batchId
+        Long batchId
 ) {}
